Fix type coversion things

+++ cis_interface/dataio/AsciiFile.h
@@ -92,7 +92,7 @@ int af_is_comment(const asciiFile_t t, const char *line) {
 static inline
 int af_readline_full(const asciiFile_t t, char **line, size_t *n) {
   if (af_is_open(t) == 1) {
-    return getline(line, n, t.fd);
+    return (int)getline(line, n, t.fd);
   }
   return -1;
 };
